requirements.txt
# 智能旅行规划助手 - Python 依赖
# 建议在 Python 3.10+ 虚拟环境中安装: pip install -r requirements.txt

# Web 框架
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# 数据校验与配置
pydantic>=2.0.0
python-dotenv>=1.0.0

# LLM 与 MCP
openai>=1.0.0
mcp>=1.0.0
fastmcp>=2.0.0

# HTTP 与工具服务
requests>=2.28.0
